pipeline/requirements.txt
# Scraping
beautifulsoup4>=4.12.0
scrapy>=2.11.0
requests>=2.31.0
urllib3>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Celery
celery[redis]>=5.3.0
redis>=5.0.0

# Config & HTTP
python-dotenv>=1.0.0
httpx>=0.25.0

# Utils
pydantic>=2.5.0
pytz>=2024.1
